Get the selected dataset

Retrieving a dataset with study-level aggregate clinical and methodological characteristics (that may act as effect modifiers) extracted from the selected systematic review of the database.

Usage

get.dataset(pmid, show.index = FALSE, show.type = FALSE)

Arguments

pmid

A scalar with the PMID number of the systematic review found in the database.

show.index

Logical to indicate whether to return a data-frame with the full name of the abbreviated characteristics. The default is FALSE (do not report).

show.type

Logical to indicate whether to return a data-frame with the type (Demographic, Clinical, Methodological) and subtype (Age, Ethnicity, Intervention, Outcome, Participant, Risk of bias, Sex, Study design, Study setting, Withdrawals) of the characteristics. The default is FALSE (do not report).

Value

get.dataset returns the following:

Dataset

A data-frame (tibble style) with rows referring to the studies and columns to the study-level aggregate characteristics as extracted from the report of the corresponding systematic review.

Characteristics_index

A data-frame (tibble style) with the rows referring to the extracted characteristics (abbreviated name) and columns to the their full name (if show.index = TRUE), their type and subtype (if show.type = TRUE).

Details

The selected dataset refers to a connected network for a specific outcome studied in the corresponding systematic review. The R package nmadb was used to retrieve the corresponding dataset. Specifically, the function readByID was employed to download the dataset in the long format. Then, the function pairwise of the R package netmeta was implemented to convert the dataset into wide format with each row repeated as many times as the number of possible comparisons made in the corresponding study. The study names (or references) and treatment comparisons, as returned by readByID, were used to locate the studies in the corresponding report of the systematic review (and Appendix, if available) and extract the characteristics available in the relevant table(s). Each characteristic occupies one column in the dataset. Characteristics pertaining to intervention features occupied one column for the experimental and another for the control treatment in the corresponding comparison.

See also

Author

Loukia M. Spineli

Examples

get.dataset(pmid = 25626481)
#> $Dataset
#> # A tibble: 6 × 36
#>   trial      treat1 treat2 arm1  arm2  continent multicenter funding sample.size
#>   <chr>       <dbl>  <dbl> <chr> <chr> <chr>     <chr>       <chr>         <dbl>
#> 1 Alberts         1      2 PLDH… carb… America   NA          yes              61
#> 2 Bafaloukos      1      3 PLDH… pacl… Europe    NA          NA              189
#> 3 CALYPSO         1      3 PLDH… pacl… internat… yes         yes             976
#> 4 Gonzalez-…      3      4 pacl… carb… Europe    NA          NA               81
#> 5 ICON4/AGO…      3      4 pacl… carb… Europe    yes         yes             802
#> 6 Pfisterer       4      5 carb… gemc… Europe    NA          yes             356
#> # ℹ 27 more variables: perc.withdrawals <dbl>, mean.treat.duration <dbl>,
#> #   SD.treat.duration <dbl>, follow.up.duration <dbl>, mean.age <dbl>,
#> #   SD.age <dbl>, mean.PFI.duration <dbl>, SD.PFI.duration <dbl>,
#> #   perc.PFI.above.365.days <dbl>, perc.prior.chemotherapy <dbl>,
#> #   up.to.2.lesions.sites <dbl>, at.least.3.lesion.sites <dbl>,
#> #   perc.serious.type <dbl>, perc.non.serious.type <dbl>, tumour.size <dbl>,
#> #   perc.measurable.disease <dbl>, perc.elevated.CA125.level <dbl>, …
#>