Dataset catalogue

A database with 217 systematic reviews with network meta-analysis (NMA) published from 2004 to April 14, 2015. The collected systematic reviews were retrieved from the archived R package nmadb and pertain to those with available data in that package.

Usage

data(index)

Format

A data-frame with 217 rows and 10 columns referring to the analysed systematic reviews with NMA and their characteristics, respectively.

Source

Papakonstantinou T. nmadb: Network Meta-Analysis Database API. R package version 1.2.0. 2019. <https://CRAN.R-project.org/package=nmadb> (archived).

Turner RM, Davey J, Clarke MJ, Thompson SG, Higgins JP. Predicting the extent of heterogeneity in meta-analysis, using empirical data from the Cochrane Database of Systematic Reviews. Int J Epidemiol 2012;41(3):818–27. doi: 10.1093/ije/dys041.

Details

The characteristics of the systematic reviews comprise:

nmadb.ID	The ID number as provided in the archived R package nmadb.

PMID	The PMID number.

First.Author	The name of the first author.

Year	The year of publication.

Journal.Name	The abbreviated name of the journal.

Title	The title of the systematic review.

Outcome.Type	The outcome type as suggested by Turner et al. (2012) and distinguished into objective, semi-objective and subjective.

Intervention.Comparison.Type	The treatment-comparator type as suggested by Turner et al. (2012) and distinguished into pharmacological versus placebo, pharmacological versus pharmacological, and non-pharmacological versus any.

Includes.ToC.where	Whether the extracted study-level aggregate characteristics were found in the main article, Appendix or both.

Source.ToC	The exact location in the systematic review where the extracted study-level aggregate characteristics were found, such as Table X in the main article, and/ or Appendix.

Comment	Notes related to the extraction for the selected systematic review(s).

Except for PMID, Includes.ToC.where, Source.ToC, and Comment, all other characteristics were retrieved from the archived R package nmadb. The database was reduced further during extraction, for instance, due to the limited reporting quality and data of the systematic review relating to the extraction process.