Get the index of specific dataset(s)

Retrieving the characteristics of one of more systematic reviews found in the database using their PMID number or year of publication.

Usage

get.dataset.index(pmid, year = NULL, show.title = FALSE, show.comment = FALSE)

Arguments

pmid

A scalar or vector with the PMID number(s) of the systematic reviews found in the database.

year

A scalar to define the year of publication. get.dataset.index will return all systematic reviews found in the database under this year.

show.title

Logical to indicate whether to show the title of the selected systematic review(s). The default is FALSE (do not show).

show.comment

Logical to indicate whether to show the comment referring to extraction notes about the selected systematic review(s). The default is FALSE (do not show).

Value

get.dataset.index returns a vector (when pmid is scalar) or a data-frame (when pmid is a vector) with the following characteristics (column names):

nmadb.ID

The ID number as provided in the R package nmadb.

PMID

The PMID number.

First.Author

The name of the first author.

Year

The year of publication.

Journal.Name

The abbreviated name of the journal.

Title

The title of the systematic review.

Outcome.Type

The outcome type as suggested by Turner et al. (2012) and distinguished into objective, semi-objective and subjective.

Intervention.Comparison.Type

The treatment-comparator type as suggested by Turner et al. (2012) and distinguished into pharmacological versus placebo, pharmacological versus pharmacological, and non-pharmacological versus any.

Includes.ToC.where

Whether the extracted study-level aggregate characteristics were found in the main article, Appendix or both.

Source.ToC

The exact location in the systematic review where the extracted study-level aggregate characteristics were found, such as Table(s) in the main article, and/ or Appendix.

Comment

Notes related to the extraction for the selected systematic review(s).

Details

When year is specified, the argument pmid should be NULL. Then, get.dataset.index returns all systematic reviews found in the database under this year.

Except for PMID, Includes.ToC.where, and Source.ToC, all other characteristics were retrieved from the R package nmadb using the function getNMADB and subsetting to those systematic reviews with available data (i.e., Data.available == "True" when using the function getNMADB). The database was reduced further during extraction, for instance, due to the reporting issues and data unavailability of the systematic review relating to the extraction process.

References

Turner RM, Davey J, Clarke MJ, Thompson SG, Higgins JP. Predicting the extent of heterogeneity in meta-analysis, using empirical data from the Cochrane Database of Systematic Reviews. Int J Epidemiol 2012;41(3):818--27. doi: 10.1093/ije/dys041.

See also

Author

Loukia M. Spineli

Examples

# For one systematic review with PMID number 25626481.
get.dataset.index(pmid = 25626481)
#>     nmadb.ID     PMID First.Author Year          Journal.Name Outcome.Type
#> 191   476343 25626481      Edwards 2015 Health Technol Assess    Objective
#>           Intervention.Comparison.Type Includes.ToC.where       Source.ToC
#> 191 pharmacological vs pharmacological Supplementary file Appendices 2 &12

# For two systematic reviews with PMID numbers 25626481 and 15147585,
# respectively.
get.dataset.index(pmid = c(25626481, 15147585))
#>     nmadb.ID     PMID First.Author Year          Journal.Name Outcome.Type
#> 1     501330 15147585        Mason 2004         BMC Fam Pract   Subjective
#> 191   476343 25626481      Edwards 2015 Health Technol Assess    Objective
#>           Intervention.Comparison.Type Includes.ToC.where            Source.ToC
#> 1           pharmacological vs placebo Supplementary file Additional File 4 & 5
#> 191 pharmacological vs pharmacological Supplementary file      Appendices 2 &12

# For systematic reviews published during 2010.
get.dataset.index(pmid = NULL, year = 2010)
#>    nmadb.ID     PMID First.Author Year               Journal.Name
#> 34   481216 20223500      Bergman 2010      Semin Arthritis Rheum
#> 39   501366 20388897        Phung 2010                       JAMA
#> 42   501430 20600036          Woo 2010           Gastroenterology
#> 33   480851 20614454        Stowe 2010 Cochrane Database Syst Rev
#> 41   501424 20638155         Wang 2010              J Hosp Infect
#> 38   501340 20713583    Middleton 2010                        BMJ
#> 36   501299 20738930      Imamura 2010      Health Technol Assess
#> 35   481571 20825624        Dakin 2010               Value Health
#> 37   501302 20828791       Jansen 2010      Semin Arthritis Rheum
#> 40   501423 20847017       Wandel 2010                        BMJ
#>      Outcome.Type Intervention.Comparison.Type Includes.ToC.where
#> 34 Semi-objective   pharmacological vs placebo Supplementary file
#> 39      Objective   pharmacological vs placebo Supplementary file
#> 42      Objective   pharmacological vs placebo          Main text
#> 33 Semi-objective   pharmacological vs placebo Supplementary file
#> 41 Semi-objective   non-pharmacological vs any          Main text
#> 38     Subjective   non-pharmacological vs any Supplementary file
#> 36     Subjective   pharmacological vs placebo               Both
#> 35      Objective   pharmacological vs placebo          Main text
#> 37 Semi-objective   pharmacological vs placebo          Main text
#> 40     Subjective   pharmacological vs placebo          Main text
#>                                         Source.ToC
#> 34                                Appendix Table 1
#> 39                                        eTable 1
#> 42                              Table 1 & Figure 3
#> 33             Characteristics of included studies
#> 41                                         Table 1
#> 38         Appendix Tables 2 (for sample size) & 3
#> 36 Table 14 (Appendix Table 10) & Appendix 9 (RoB)
#> 35                                         Table 2
#> 37                                         Table 1
#> 40                   Table 1 & 2 (for sample size)